requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
requests==2.31.0
psycopg2-binary==2.9.9
